+++ dio/lib/src/adapters/browser_adapter.dart
@@ -73,18 +73,26 @@ class BrowserHttpClientAdapter implements HttpClientAdapter {
         connectionTimeout,
         () {
           if (!completer.isCompleted) {
+            xhr.abort();
             completer.completeError(
-              DioError(
+              DioError.connectionTimeout(
                 requestOptions: options,
-                error: 'Connecting timed out [${options.connectTimeout}ms]',
-                type: DioErrorType.connectTimeout,
+                timeout: connectionTimeout,
               ),
               StackTrace.current,
             );
-            xhr.abort();
+            return;
           } else {
             // connectTimeout is triggered after the fetch has been completed.
           }
+          xhr.abort();
+          completer.completeError(
+            DioError.connectionTimeout(
+              requestOptions: options,
+              timeout: options.connectTimeout!,
+            ),
+            StackTrace.current,
+          );
         },
       );
     }
@@ -107,10 +115,9 @@ class BrowserHttpClientAdapter implements HttpClientAdapter {
         if (duration > sendTimeout) {
           uploadStopwatch.stop();
           completer.completeError(
-            DioError(
+            DioError.sendTimeout(
+              timeout: sendTimeout,
               requestOptions: options,
-              error: 'Sending timed out [${options.sendTimeout}ms]',
-              type: DioErrorType.sendTimeout,
             ),
             StackTrace.current,
           );
@@ -141,10 +148,9 @@ class BrowserHttpClientAdapter implements HttpClientAdapter {
         if (duration > reveiveTimeout) {
           downloadStopwatch.stop();
           completer.completeError(
-            DioError(
+            DioError.receiveTimeout(
+              timeout: options.receiveTimeout!,
               requestOptions: options,
-              error: 'Receiving timed out [${options.receiveTimeout}ms]',
-              type: DioErrorType.receiveTimeout,
             ),
             StackTrace.current,
           );
@@ -162,11 +168,12 @@ class BrowserHttpClientAdapter implements HttpClientAdapter {
       connectTimeoutTimer?.cancel();
       // Unfortunately, the underlying XMLHttpRequest API doesn't expose any
       // specific information about the error itself.
+      // See also: https://developer.mozilla.org/en-US/docs/Web/API/XMLHttpRequestEventTarget/onerror
       completer.completeError(
-        DioError(
-          type: DioErrorType.response,
-          error: 'XMLHttpRequest error.',
+        DioError.connectionError(
           requestOptions: options,
+          reason: 'The XMLHttpRequest onError callback was called. '
+              'This typically indicates an error on the network layer.',
         ),
         StackTrace.current,
       );