+++ dio/lib/src/dio_error.dart
@@ -2,8 +2,8 @@ import 'options.dart';
 import 'response.dart';
 
 enum DioErrorType {
-  /// It occurs when url is opened timeout.
-  connectTimeout,
+  /// Caused by a connection timeout.
+  connectionTimeout,
 
   /// It occurs when url is sent timeout.
   sendTimeout,
@@ -11,25 +11,99 @@ enum DioErrorType {
   ///It occurs when receiving timeout.
   receiveTimeout,
 
-  /// When the server response, but with a incorrect status, such as 404, 503...
-  response,
+  /// The [DioError] was caused by an incorrect status code as configured by
+  /// [ValidateStatus].
+  badResponse,
 
   /// When the request is cancelled, dio will throw a error with this type.
   cancel,
 
+  /// Caused for example by a `xhr.onError` or SocketExceptions.
+  connectionError,
+
   /// Default error type, Some other Error. In this case, you can
   /// use the DioError.error if it is not null.
-  other,
+  unknown,
 }
 
-/// DioError describes the error info  when request failed.
+extension _DioErrorTypeExtension on DioErrorType {
+  String toPrettyDescription() {
+    switch (this) {
+      case DioErrorType.connectionTimeout:
+        return 'connection timeout';
+      case DioErrorType.sendTimeout:
+        return 'send timeout';
+      case DioErrorType.receiveTimeout:
+        return 'receive timeout';
+      case DioErrorType.badResponse:
+        return 'bad response';
+      case DioErrorType.cancel:
+        return 'request cancelled';
+      case DioErrorType.connectionError:
+        return 'connection error';
+      case DioErrorType.unknown:
+        return 'unknown';
+    }
+  }
+}
+
+/// DioError describes the exception info when a request failed.
 class DioError implements Exception {
+  /// Prefer using one of the other constructors.
+  /// They're most likely better fitting.
   DioError({
     required this.requestOptions,
     this.response,
-    this.type = DioErrorType.other,
+    this.type = DioErrorType.unknown,
     this.error,
+    this.stackTrace,
+    this.message,
   });
+
+  DioError.badResponse({
+    required int statusCode,
+    required this.requestOptions,
+    required this.response,
+  })  : type = DioErrorType.badResponse,
+        message = 'The request returned an '
+            'invalid status code of $statusCode.';
+
+  DioError.connectionTimeout({
+    required Duration timeout,
+    required this.requestOptions,
+    this.error,
+    this.stackTrace,
+  })  : type = DioErrorType.connectionTimeout,
+        message = 'The request connection took '
+            'longer than $timeout. It was aborted.';
+
+  DioError.sendTimeout({
+    required Duration timeout,
+    required this.requestOptions,
+  })  : type = DioErrorType.sendTimeout,
+        message = 'The request took '
+            'longer than $timeout to send data. It was aborted.';
+
+  DioError.receiveTimeout({
+    required Duration timeout,
+    required this.requestOptions,
+  })  : type = DioErrorType.receiveTimeout,
+        message = 'The request took '
+            'longer than $timeout to receive data. It was aborted.';
+
+  DioError.requestCancelled({
+    required this.requestOptions,
+    required Object? reason,
+    this.stackTrace,
+  })  : type = DioErrorType.cancel,
+        message = 'The request was cancelled.',
+        error = reason;
+
+  DioError.connectionError({
+    required this.requestOptions,
+    required String reason,
+  })  : type = DioErrorType.connectionError,
+        message = 'The connection errored: $reason';
 
   /// Request info.
   RequestOptions requestOptions;
@@ -42,24 +116,20 @@ class DioError implements Exception {
 
   /// The original error/exception object; It's usually not null when `type`
   /// is DioErrorType.other
-  dynamic error;
+  Object? error;
 
-  StackTrace? _stackTrace;
+  /// The stacktrace of the original error/exception object;
+  /// It's usually not null when `type` is DioErrorType.other
+  StackTrace? stackTrace;
 
-  set stackTrace(StackTrace? stack) => _stackTrace = stack;
-
-  StackTrace? get stackTrace => _stackTrace;
-
-  String get message => (error?.toString() ?? '');
+  String? message;
 
   @override
   String toString() {
-    var msg = 'DioError [$type]: $message';
-    if (error is Error) {
-      msg += '\n${(error as Error).stackTrace}';
-    }
-    if (_stackTrace != null) {
-      msg += '\nSource stack:\n$stackTrace';
+    var msg = 'DioError [${type.toPrettyDescription()}]: $message';
+    final error = this.error;
+    if (error != null) {
+      msg += '\nError: $error';
     }
     return msg;
   }